move view provider into adequate module
